% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_content.R
\name{clean_content}
\alias{clean_content}
\title{Clean scraped content}
\usage{
clean_content(df)
}
\arguments{
\item{df}{Data frame with scraped content. Output of \code{"get_content"}.}
}
\value{
Data frame with cleaned content of scraped rules.
}
\description{
Clean scraped content
}
\examples{
\dontrun{
chapter_content <- get_content("http://www.prarulebook.co.uk/rulebook/Content/Chapter/242047/16-11-2007")
clean_content(chapter_content)
}
}
