% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_layer.R
\name{recode_layer}
\alias{recode_layer}
\title{Recode layer of the rulebook}
\usage{
recode_layer(from, to = "all", structure_file)
}
\arguments{
\item{from}{Source layer. Rule, chapter, part, sector.}

\item{to}{Target layer. Rule, chapter, part, sector.}

\item{structure_file}{Data frame. Output of \code{"get_structure"}.}
}
\value{
Data frame with recoded layers.
}
\description{
Recode layer of the rulebook
}
\examples{
\dontrun{
rulebook_structure <- get_structure("16-11-2017")
recode_layer("http://www.prarulebook.co.uk/rulebook/Content/Chapter/211140/16-11-2017",
to = "all", structure_file = rulebook_structure)
}
}
