% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_menu.R
\name{scrape_menu}
\alias{scrape_menu}
\title{Scrape the menu}
\usage{
scrape_menu(url, selector, structure_type = "both")
}
\arguments{
\item{url}{Full URL to scrape}

\item{selector}{CSS selector to scrape. Use Chrome with SelectorGadget to find the relevant selector.}

\item{structure_type}{Type of structure to scrape: text, url, both.}
}
\value{
Data frame with text, url, or both.
}
\description{
Scrapes names and URLs from the PRA rulebook menu. The menu is on the left-hand side of the rulebook website.
}
\examples{
\dontrun{
scrape_menu("http://www.prarulebook.co.uk/rulebook/Home/Handbook/16-11-2007", ".nav-child a")
scrape_menu("http://www.prarulebook.co.uk/rulebook/Home/Handbook/16-11-2007", ".nav-child a", "text")
}
}
