% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_menu_rule.R
\name{scrape_menu_rule}
\alias{scrape_menu_rule}
\title{Scrape rules within the scrape_menu function}
\usage{
scrape_menu_rule(url, nodes_only, rulebook_date)
}
\arguments{
\item{rulebook_date}{}
}
\value{
Data frame
}
\description{
Helper function
}
\examples{
\dontrun{
has_rules_url <- "http://www.prarulebook.co.uk/rulebook/Content/Chapter/218624/22-03-2006"
no_rules_url <- "http://www.prarulebook.co.uk/rulebook/Content/Chapter/219425/22-03-2006"
nodes_only <- httr::GET(no_rules_url) \%>\% extract_results()
nodes_only <-
nodes_only \%>\%
  xml2::read_html() \%>\%
  rvest::html_nodes("a")
scrape_menu_rule()
}
}
