% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_part_structure.R
\name{scrape_part_structure}
\alias{scrape_part_structure}
\title{Scrape part structure}
\usage{
scrape_part_structure()
}
\value{
Data frame with scraped part structure.
}
\description{
Scrape part structure
}
