% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rule_structure.R
\name{scrape_rule_structure}
\alias{scrape_rule_structure}
\title{Scrape rule structure}
\usage{
scrape_rule_structure(df)
}
\arguments{
\item{df}{Data frame with chapter URLs. Output of \code{"scrape_chapter_structure"}.}
}
\value{
Data frame with scraped rule structure.
}
\description{
Scrape rule structure
}
\examples{
\dontrun{
sectors <- scrape_sector_structure("http://www.prarulebook.co.uk/rulebook/Home/Handbook/16-11-2007")
parts <- scrape_part_structure(sectors)
chapters <- scrape_chapter_structure(parts)
scrape_rule_structure(chapters)
}
}
