% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sector_structure.R
\name{scrape_sector_structure}
\alias{scrape_sector_structure}
\title{Scrape sector structure}
\usage{
scrape_sector_structure(x)
}
\arguments{
\item{x}{String. Top URL.}
}
\value{
Data frame with scraped sector structure.
}
\description{
Scrape sector structure
}
\examples{
\dontrun{
scrape_sector_structure("http://www.prarulebook.co.uk/rulebook/Home/Handbook/16-11-2007")
}
}
