% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_content}
\alias{get_content}
\title{Scrape the Rulebook content}
\usage{
get_content(x, type = "text", single_rule_selector = NULL)
}
\arguments{
\item{x}{String. URL to scrape.}

\item{type}{String. Type of information to scrape. "Text" or "links".}

\item{single_rule_selector}{String. Optional. CSS selector for individual rules.}
}
\value{
Data frame with URLs and corresponding text.
}
\description{
Extract the full text or links from the PRA Rulebook given the URL.
}
\examples{
\dontrun{
get_content("http://www.prarulebook.co.uk/rulebook/Content/Chapter/242047/16-11-2007")
get_content("http://www.prarulebook.co.uk/rulebook/Content/Chapter/242047/16-11-2007", "links")
get_content("http://www.prarulebook.co.uk/rulebook/Content/Rule/216149/16-11-2007#216149",
"links", "yes")
}
}
