% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_part_structure.R
\name{scrape_part_structure}
\alias{scrape_part_structure}
\title{Scrape part structure}
\usage{
scrape_part_structure(df)
}
\arguments{
\item{df}{Data frame with sector URLs. Output of \code{"scrape_sector_structure"}.}
}
\value{
Data frame with scraped part structure.
}
\description{
Scrape part structure
}
\examples{
\dontrun{
sectors <- scrape_sector_structure("http://www.prarulebook.co.uk/rulebook/Home/Handbook/16-11-2007")
scrape_part_structure(sectors)
}
}
