% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rule_id.R
\name{scrape_rule_id}
\alias{scrape_rule_id}
\title{Scrape rule ID}
\usage{
scrape_rule_id(rule_urls, type = "text")
}
\arguments{
\item{rule_urls}{Data frame with rule url. Output of \code{"scrape_rule_structure"}.}

\item{type}{String. Type of data to scrape: text or links.}
}
\value{
Data frame with rule text and links.
}
\description{
Scrape rule ID
}
\examples{
\dontrun{
chapters_df <- get_structure("16-11-2007", layer = "chapter")
rules12 <- scrape_rule_structure(chapters_df[1:2,], "16-11-2007")
scrape_rule_id(rules12[1,])
}
}
