% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rule_id.R
\name{scrape_rule_id}
\alias{scrape_rule_id}
\title{Scrape rule ID}
\usage{
scrape_rule_id(url, selector_rule_no, selector_rule_text)
}
\arguments{
\item{url}{String. URL to scrape. Output from \code{"scrape_rule_structure"}.}

\item{selector_rule_no}{Rule number CSS selector}

\item{selector_rule_text}{Rule text CSS selector}
}
\value{
Data frame with rule text and links.
}
\description{
Scrape rule ID
}
\examples{
\dontrun{
chapters_df <- get_structure("16-11-2007", layer = "chapter")
rules12 <- scrape_rule_structure(chapters_df[1:2,], "16-11-2007")
scrape_rule_id(rules12$rule_url[1], rules12$rule_number_sel[1], rules12$rule_text_sel[1])
}
}
